requirements.txt
streamlit
numpy
matplotlib
plotly
google-generativeai
python-dotenv
ipython
streamlit-lottie
opencv-python